+++ types/src/transaction/transaction_v1/transaction_v1_builder.rs
@@ -8,7 +8,7 @@ use rand::Rng;
 use super::{
     super::{
         InitiatorAddr, TransactionEntryPoint, TransactionInvocationTarget, TransactionRuntime,
-        TransactionScheduling, TransactionSessionKind, TransactionTarget,
+        TransactionScheduling, TransactionTarget,
     },
     transaction_v1_body::arg_handling,
     InitiatorAddrAndSecretKey, PricingMode, TransactionV1, TransactionV1Body,
@@ -286,9 +286,8 @@ impl<'a> TransactionV1Builder<'a> {
 
     /// Returns a new `TransactionV1Builder` suitable for building a transaction for running session
     /// logic, i.e. compiled Wasm.
-    pub fn new_session(kind: TransactionSessionKind, module_bytes: Bytes) -> Self {
+    pub fn new_session(category: TransactionCategory, module_bytes: Bytes) -> Self {
         let target = TransactionTarget::Session {
-            kind,
             module_bytes,
             runtime: Self::DEFAULT_RUNTIME,
         };
@@ -296,7 +295,7 @@ impl<'a> TransactionV1Builder<'a> {
             RuntimeArgs::new(),
             target,
             TransactionEntryPoint::Call,
-            TransactionCategory::Large as u8,
+            category as u8,
             Self::DEFAULT_SCHEDULING,
         );
         TransactionV1Builder::new(body)