+++ node/CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.  The format
 ### Removed
 * Remove the JSON-RPC and speculative execution interfaces.
 * Remove chainspec setting `highway.performance_meter.blocks_to_consider` and the entire `highway.performance_meter` section.
+* Remove chainspec setting `highway.reduced_reward_multiplier`